[schedule] Add handling of verbatim sessions
Verbatim sessions are those for which no lesson file is written.
Examples include review lessons and TBD lessons. They are distinguished
by having the id 'uniq_id@title', where uniq_id is a unique string that
distinguishes the session from other sessions with the same title, and
title is a human-readable title for the session.

E.g.
    "0@Review": Jan 1 2018