[subclubs] Change behaviour of blank schedule entries
Blank schedule entries now mean that the lesson date is the same
as the previous lesson, not that the value of the entry is the
same as the previous lesson

e.g.:

lesson-one: Jan 1 2018
lesson-two: 7
lesson-three:

used to mean that lesson-two was on Jan 8 2018, and lesson-three
on Jan 15 2018 (repeating the offset). It now means that both
lesson-two and lesson-three are on Jan 8 2018.